fodasse tou cansado

+++ Trabalhador.cs
@@ -15,7 +15,7 @@ namespace ProjectodeDA
     public partial class Trabalhador : Pessoa
     {
         public int IdRestaurante { get; set; }
-        public double Salario { get; set; }
+        public decimal Salario { get; set; }
         public string Posicao { get; set; }
     
         public virtual Restaurante Restaurante { get; set; }
